+++ packages/terra-popup/src/Popup.jsx
@@ -216,7 +216,7 @@ class Popup extends React.Component {
     }
   }
 
-  createPopupContent(boundingFrame, showArrow) {
+  createPopupContent(boundingFrame, showArrow, hookshotContentPosition) {
     const boundsProps = {
       contentHeight: PopupHeights[this.props.contentHeight] || PopupHeights['80'],
       contentWidth: PopupWidths[this.props.contentWidth] || PopupWidths['240'],
@@ -257,6 +257,7 @@ class Popup extends React.Component {
         isWidthAutomatic={this.props.contentWidth === 'auto'}
         isFocusedDisabled={this.props.isContentFocusDisabled}
         onClick={this.props.onClick}
+        hookshotContentPosition={hookshotContentPosition}
       >
         {this.props.children}
       </PopupContent>
@@ -282,6 +283,7 @@ class Popup extends React.Component {
       onRequestClose,
       targetRef,
       targetAttachment,
+      ...customProps
     } = this.props;
     /* eslint-enable no-unused-vars */
 
@@ -302,7 +304,7 @@ class Popup extends React.Component {
     if (showArrow) {
       cOffset = PopupUtils.getContentOffset(cAttachment, tAttachment, this.props.targetRef(), PopupArrow.Opts.arrowSize, cornerSize);
     }
-    const hookshotContent = this.createPopupContent(boundingRef ? boundingRef() : undefined, showArrow);
+    const hookshotContent = this.createPopupContent(boundingRef ? boundingRef() : undefined, showArrow, customProps.hookshotContentPosition);
 
     return (
       <React.Fragment>