+++ packages/terra-hookshot/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Implemented dynamic content positioning in HookshotContent component: Overrides content position with `hookshotContentPosition` from customProps or defaults to 'absolute' if not provided.
+
 ## 5.41.0 - (July 31, 2023)
 
 * Changed